+++ src/modules/predefinedDimensions.js
@@ -8,6 +8,7 @@ export const DIMENSION_ID_DATA = 'dx'
 export const DIMENSION_ID_PERIOD = 'pe'
 export const DIMENSION_ID_ORGUNIT = 'ou'
 export const DIMENSION_ID_ASSIGNED_CATEGORIES = 'co'
+export const DIMENSION_PROP_NO_ITEMS = 'noItems'
 
 const FIXED_DIMENSIONS = {
     [DIMENSION_ID_DATA]: {
@@ -28,23 +29,36 @@ const FIXED_DIMENSIONS = {
         iconName: 'OrgUnitIcon',
         icon: OrgUnitIcon,
     },
+}
+
+const DYNAMIC_DIMENSIONS = {
     [DIMENSION_ID_ASSIGNED_CATEGORIES]: {
         id: DIMENSION_ID_ASSIGNED_CATEGORIES,
         name: i18n.t('Assigned Categories'),
         iconName: 'AssignedCategoriesIcon',
         icon: AssignedCategoriesIcon,
-        noItems: true,
+        [DIMENSION_PROP_NO_ITEMS]: true,
     },
 }
 
-export const filterOutFixedDimensions = dimensionIds =>
+const PREDEFINED_DIMENSIONS = {
+    ...FIXED_DIMENSIONS,
+    ...DYNAMIC_DIMENSIONS,
+}
+
+export const filterOutPredefinedDimensions = dimensionIds =>
     dimensionIds.filter(
-        dimensionId => !Object.keys(FIXED_DIMENSIONS).includes(dimensionId)
+        dimensionId => !Object.keys(PREDEFINED_DIMENSIONS).includes(dimensionId)
     )
 
-export const getFixedDimensionProp = (dimensionId, propName) =>
-    (FIXED_DIMENSIONS[dimensionId] || {})[propName]
+export const getPredefinedDimensionProp = (dimensionId, propName) =>
+    (PREDEFINED_DIMENSIONS[dimensionId] || {})[propName]
 
-export const getDimensionById = dimensionId => FIXED_DIMENSIONS[dimensionId]
+export const getDimensionById = dimensionId =>
+    PREDEFINED_DIMENSIONS[dimensionId]
+
+export const getPredefinedDimensions = () => PREDEFINED_DIMENSIONS
 
 export const getFixedDimensions = () => FIXED_DIMENSIONS
+
+export const getDynamicDimensions = () => DYNAMIC_DIMENSIONS