+++ src/modules/visTypeToLayoutType.js
@@ -17,6 +17,7 @@ import {
     LAYOUT_TYPE_DEFAULT,
     LAYOUT_TYPE_PIE,
     LAYOUT_TYPE_YEAR_OVER_YEAR,
+    LAYOUT_TYPE_PIVOT_TABLE,
 } from './layoutTypes'
 
 const visTypeToLayoutType = {
@@ -32,7 +33,7 @@ const visTypeToLayoutType = {
     [VIS_TYPE_YEAR_OVER_YEAR_LINE]: LAYOUT_TYPE_YEAR_OVER_YEAR,
     [VIS_TYPE_YEAR_OVER_YEAR_COLUMN]: LAYOUT_TYPE_YEAR_OVER_YEAR,
     [VIS_TYPE_SINGLE_VALUE]: LAYOUT_TYPE_PIE,
-    [VIS_TYPE_PIVOT_TABLE]: LAYOUT_TYPE_DEFAULT,
+    [VIS_TYPE_PIVOT_TABLE]: LAYOUT_TYPE_PIVOT_TABLE,
 }
 
 export const getLayoutTypeByVisType = visType => visTypeToLayoutType[visType]